lesson-6: packages refactoring
