+++ app/src/main/java/com/rogoz208/notesv2/data/App.kt
@@ -3,6 +3,10 @@ package com.rogoz208.notesv2.data
 import android.app.Application
 import android.content.Context
 import android.content.IntentFilter
+import com.rogoz208.notesv2.data.log.MyAnalytics
+import com.rogoz208.notesv2.data.log.NetworkStateReceiver
+import com.rogoz208.notesv2.data.repos.NotesRepoImpl
+import com.rogoz208.notesv2.data.repos.UrlPreviewRepoImpl
 import com.rogoz208.notesv2.domain.entities.NoteEntity
 import com.rogoz208.notesv2.domain.repos.NotesRepo
 import com.rogoz208.notesv2.domain.repos.UrlPreviewRepo
