+++ app/src/main/java/com/rogoz208/notesv2/data/log/NetworkStateReceiver.kt
@@ -1,4 +1,4 @@
-package com.rogoz208.notesv2.data
+package com.rogoz208.notesv2.data.log
 
 import android.content.BroadcastReceiver
 import android.content.Context
