+++ app/src/main/java/com/rogoz208/notesv2/domain/entities/NoteEntity.kt
@@ -9,3 +9,5 @@ data class NoteEntity(
     var detail: String,
     var creationDate: Long?
 ) : Parcelable
+
+// TODO: var to val (lesson-8 - 00:45:17)
